update admin,manager&player permission

+++ src/main/java/iob/controllers/AdminController.java
@@ -11,17 +11,22 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import iob.boundries.*;
+import iob.data.UserEntity;
+import iob.data.UserRole;
 import iob.logic.ActivitiesService;
+import iob.logic.ExtendedUserService;
 import iob.logic.InstancesService;
+import iob.logic.UserNotFoundException;
 import iob.logic.UsersService;
 
 
 @RestController
 public class AdminController {
-	private UsersService admin;
+	private ExtendedUserService admin;
 	private InstancesService instancesService;
 	private ActivitiesService activitiesService;
 	
@@ -34,7 +39,7 @@ public class AdminController {
 		this.instancesService = instancesService;
 	}
 	@Autowired
-	public void setUsersService(UsersService admin) {
+	public void setUsersService(ExtendedUserService admin) {
 		this.admin = admin;
 	}
 	
@@ -58,8 +63,12 @@ public class AdminController {
 	@RequestMapping(
 			method = RequestMethod.DELETE,
 			path = "/iob/admin/users")
-	public void deleteAllUsers () {
-		//TODO: validate admin permission
+	public void deleteAllUsers (
+			@RequestParam(name="userDomain", required = true) String domain,
+			@RequestParam(name="userEmail", required = true) String email
+			) {
+		// Get user data from DB and check if ADMIN
+		this.admin.checkUserPermission((new UserID(domain, email)).toString(), UserRole.ADMIN);
 		admin.deleteAllUsers();
 	}
 	
@@ -67,16 +76,25 @@ public class AdminController {
 	@RequestMapping(
 			method = RequestMethod.DELETE,
 			path = "/iob/admin/instances")
-		public void deleteAllInstances () {
-			//TODO: validate admin permission
+		public void deleteAllInstances (
+				@RequestParam(name="userDomain", required = true) String domain,
+				@RequestParam(name="userEmail", required = true) String email
+				) {
+			// Get user data from DB and check if ADMIN
+			this.admin.checkUserPermission((new UserID(domain, email)).toString(), UserRole.ADMIN);
 			instancesService.deleteAllInstances();
 		}
 	
 	@RequestMapping(
 			method = RequestMethod.DELETE,
 			path = "/iob/admin/activities")
-	public void deleteAllActivities () {
-		//TODO: validate admin permission
+	public void deleteAllActivities (
+			@RequestParam(name="userDomain", required = true) String domain,
+			@RequestParam(name="userEmail", required = true) String email
+			) {
+		// Get user data from DB and check if ADMIN
+		this.admin.checkUserPermission((new UserID(domain, email)).toString(), UserRole.ADMIN);
 		activitiesService.deleteAllActivities();
 	}
+	
 }
